Increased robustness

+++ premiumize-cli-downloader.sh
@@ -2,10 +2,10 @@
 
 source "/opt/premiumize.me-cli-downloader/premiumize-cli-downloader.conf"
 
-BOUNDARY="---------------------------312412633113176"
-TEMP_FILE="temp.file"
-LINKS_FILE="temp.links"
 DLC_FILE=$1
+BOUNDARY="---------------------------312412633113176"
+TEMP_FILE=".premiumize.file"
+LINKS_FILE=".premiumize.links"
 SEED="2or48h"
 
 if [ ! -z $DEFAULT_DOWNLOAD_LOCATION ] ; then
@@ -61,7 +61,7 @@ curl -s "https://www.premiumize.me/api/transfer/create" \
             --data-binary @$TEMP_FILE | \
 jq -r -c '.content[]' | \
 while read -r line; do 
-    if [[ $line == "http://ul.to"* ]] ; then
+    if [[ $line == "http://ul.to"* ||  $line == "http://uploaded.net"* ]] ; then
         echo "  Getting premium link for ${line}..."
         curl -s "https://www.premiumize.me/api/transfer/create" \
                     -H "Host: www.premiumize.me" \
@@ -76,29 +76,53 @@ while read -r line; do
     fi
 done  
 
-echo "Getting file names and downloading files..."
-rm $TEMP_FILE
-while read -r url ; do
-    FILENAME=$(echo $url | sed -e 's/^.*&f=//g')
-    echo $FILENAME >> $TEMP_FILE
-    echo "  Downloading file ${FILENAME}..."
-    curl $url -o $FILENAME -#
-done < "${LINKS_FILE}"
+#
+# Iterating over links file (if it exists), downloading each file and extracting them
+# Todo: Spawn curl process with `&` and wait for them to finish
+#
+if [ ! -e $LINKS_FILE ] ; then
+    echo "Unable to retrieve premium links!"
+    exit
+else 
+    rm $TEMP_FILE
+    echo "Getting file names and downloading files..."
+    while read -r url ; do
+        FILENAME=$(echo $url | sed -e 's/^.*&f=//g')
+        echo $FILENAME >> $TEMP_FILE
+        echo "  Downloading file ${FILENAME}..."
+        curl $url -o $FILENAME -#
+    done < "${LINKS_FILE}"
 
-echo "Trying to extract files..."
-if [[ $FILENAME == *".rar" ]] ; then
-    echo "  Archive is rar, extracting..."
-    unrar e -o+  $FILENAME
+    if [ ! -e $FILENAME ] ; then
+        echo "$FILENAME does not exist, unable to extract"
+        exit
+    else
+        echo "Trying to extract files..."
+        if [[ $FILENAME == *".rar" ]] ; then
+            echo "  Archive is rar, extracting..."
+            unrar e -o+  $FILENAME
+        fi
+    fi
 fi
 
 echo "Finished, just cleaning up..."
-while read -r file ; do
-    echo "  Removing $file"
-    rm $file
-done < "${TEMP_FILE}"
+if [ -e $TEMP_FILE ] ; then
+    while read -r file ; do
+        if [ -e $file ] ; then
+            echo "  Removing $file"
+            rm $file
+        else
+            echo "$file does not exist, unable to delete"
+        fi
+    done < "${TEMP_FILE}"
+else
+    echo "$TEMP_FILE does not exist!"
+fi
 
-echo "  Removing DLC file $DLC_FILE"
-rm $DLC_FILE
+if [ -e $DLC_FILE ] ; then
+    echo "  Removing DLC file $DLC_FILE"
+    rm $DLC_FILE
+fi
 
 if [ -e $TEMP_FILE ] ; then
     echo "  Removing temp file $TEMP_FILE"
